requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
bcrypt==4.2.1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
distro==1.9.0
ecdsa==0.19.0
elastic-transport==8.15.1
elasticsearch==8.17.0
exceptiongroup==1.2.2
fastapi==0.115.4
geographiclib==2.0
geopy==2.4.1
googletrans-py39==4.0.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.28.1
humanfriendly==10.0
hyperframe==6.0.1
idna==3.10
jiter==0.8.2
openai==1.58.1
passlib==1.7.4
pillow==11.0.0
pyasn1==0.6.1
pydantic==2.9.2
pydantic-settings==2.7.0
pydantic_core==2.23.4
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
requests==2.32.3
rsa==4.9
six==1.17.0
sniffio==1.3.1
starlette==0.41.2
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3

